+++ src/supy/_load.py
@@ -253,8 +253,7 @@ df_info_suews_cal_multitsteps = gen_suews_arg_info_df(
 df_var_info = df_info_suews_cal_multitsteps.merge(
     df_info_suews_cal_main,
     how="outer",
-)
-df_var_info = df_var_info.set_index("name")
+).set_index("name")
 
 
 # load model settings
@@ -276,8 +275,6 @@ def load_SUEWS_nml_simple(path_file):
 
 @functools.lru_cache(maxsize=128)
 def load_SUEWS_nml(p_nml):
-    # remove case issues
-    # xfile = path_insensitive(xfile)
     try:
         p_nml = Path(path_insensitive(str(p_nml))).resolve()
         parser = f90nml.Parser()
@@ -304,8 +301,6 @@ def load_SUEWS_table(path_file):
     except FileNotFoundError:
         logger_supy.exception(f"{path_file} does not exists!")
     else:
-        # fileX = path_insensitive(fileX)
-        # str_file = str(path_file)
         try:
             rawdata = pd.read_csv(
                 path_file,
@@ -317,7 +312,6 @@ def load_SUEWS_table(path_file):
                 index_col=0,
             )
             rawdata = rawdata.dropna()
-            # rawdata = rawdata.apply(pd.to_numeric)
             rawdata.index = rawdata.index.astype(int)
             return rawdata
         except Exception as err:
@@ -1318,7 +1312,7 @@ dict_RunControl_default = {
     "disaggmethod": 1,
     "disaggmethodestm": 1,
     "raindisaggmethod": 100,
-    "localclimatemethod": 0, # 0: use forcing data; 1: use local climate - modelled
+    "localclimatemethod": 0,  # 0: use forcing data; 1: use local climate - modelled
     "rainamongn": -999,
     "multrainamongn": -999,
     "multrainamongnupperi": -999,
@@ -1590,11 +1584,9 @@ def load_nml_multi(fn_nml):
 def add_file_gridlayout_df(df_init):
     # load all nml info from file names:('file_init', '0')
     dict_grid_layout = (
-        df_init[("file_gridlayout", "0")]
-        .map(lambda fn: load_SUEWS_nml(fn))
-        .to_dict()
+        df_init[("file_gridlayout", "0")].map(lambda fn: load_SUEWS_nml(fn)).to_dict()
     )
-    df_grid_layout=pd.DataFrame.from_dict(dict_grid_layout, orient="index")
+    df_grid_layout = pd.DataFrame.from_dict(dict_grid_layout, orient="index")
     df_grid_layout.index.set_names(["grid"], inplace=True)
 
     # copy column names from df_init
@@ -1884,11 +1876,11 @@ def load_InitialCond_grid_df(path_runcontrol, force_reload=True):
     logger_supy.debug("adding initial temperatures...")
     df_init = add_temp_init_df(df_init)
     # if 'localclimatemethod' in df_init.columns:
-        # print()
-        # print('7')
-        # print('localclimatemethod is in df_init')
+    # print()
+    # print('7')
+    # print('localclimatemethod is in df_init')
     # else:
-        # print('localclimatemethod is not in df_init')
+    # print('localclimatemethod is not in df_init')
 
     # sort column names for consistency
     logger_supy.debug("setting grid level...")
@@ -1900,9 +1892,9 @@ def load_InitialCond_grid_df(path_runcontrol, force_reload=True):
     # print()
     # print('8')
     # if 'localclimatemethod' in df_init.columns:
-        # print('localclimatemethod is in df_init')
+    # print('localclimatemethod is in df_init')
     # else:
-        # print('localclimatemethod is not in df_init')
+    # print('localclimatemethod is not in df_init')
 
     # normalise surface fractions to prevent non-1 sums
     df_sfr_surf = df_init.sfr_surf.copy()